INTEL/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.57.0
django-allauth-2fa==0.11.1
django-axes==6.1.1
django-otp==1.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
mysqlclient==2.2.0
oauthlib==3.2.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pypng==0.20220715.0
python3-openid==3.2.0
pytz==2023.3.post1
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.6
